[MA-080125] Use enum for bottom sheet display
